src/main/java/com/playkuround/playkuroundserver/domain/user/application/UserLogoutService.java
package com.playkuround.playkuroundserver.domain.user.application;

import com.playkuround.playkuroundserver.domain.auth.token.application.TokenService;
import com.playkuround.playkuroundserver.domain.user.domain.User;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@RequiredArgsConstructor
public class UserLogoutService {

    private final TokenService tokenService;

    @Transactional
    public void logout(User user) {
        tokenService.deleteRefreshTokenByUser(user);
    }

}
